continued reorganizing validation logic, began adding documentation to the README about using the module
