+++ index.js
@@ -1,5 +1,5 @@
 // import validations 
-var validate = require('./validate');
+var validate = require('./validators');
 
 // export validations for use in other applications
 module.exports = validate;
